requirements.txt
# Requirements for the Django Web API
djangorestframework==3.13.1
# Production ASGI server
daphne==3.0.2
psycopg2-binary==2.9.3
pandas==1.4.2
djangorestframework-gis==1.0
django-filter==21.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.21.1
drf-nested-routers==0.93.4